requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.22.0
backoff==2.2.1
bcrypt==4.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
blinker==1.4
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
clarabel==0.9.0
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1

contourpy==1.2.1
cramjam==2.8.3
cryptography==3.4.8
cvxpy==1.5.3
cycler==0.12.1
dask==2024.8.1
dask-expr==1.1.11
dataclasses-json==0.6.7
datashaper==0.0.49
dbus-python==1.2.18
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
devtools==0.12.2
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.7.0

Django==5.1
document==1.0
docx2txt==0.8
ecos==2.0.14
emoji==2.12.1
environs==11.0.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.112.0
fastparquet==2024.5.0
fasttext==0.9.3
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gdown==5.2.0
gemini-api==0.1.6
geneimpacts==0.3.7
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
graphrag==0.3.2
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyppo==0.4.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.4
inheritance==0.1.5
interchange==2021.0.4
ipython==8.26.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
json_repair==0.26.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.2.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.11.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.32
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.99
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linkify-it-py==2.0.3
llama-cloud==0.0.15
llama-index==0.11.2
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.2
llama-index-embeddings-openai==0.2.3
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.0
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.5.0
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.8.4
neo4j==5.23.1
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.40.8
openpyxl==1.8.6
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.7
osqp==0.6.7.post1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pansi==2020.7.3
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
portalocker==2.10.1
posthog==3.5.2
POT==0.9.4
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py2neo==2021.2.4
pyaml-env==1.2.1
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pylance==0.15.0
pyngrok==7.2.0
pynndescent==0.5.13
pyparsing==2.4.7
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.2
qdldl==0.1.7.post4
rapidfuzz==3.9.6
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.12.0
scs==3.2.7
seaborn==0.13.2
SecretStorage==3.3.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.38.0
striprtf==0.0.26
swifter==1.4.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
textual==0.76.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
uc-micro-py==1.0.3
ufw==0.36.1
umap-learn==0.5.6
unattended-upgrades==0.1
unstructured==0.15.7
unstructured-client==0.25.5
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.20.0
wadllib==1.3.6
watchdog==4.0.2
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
wrapt==1.16.0
yachalk==0.1.6
yarl==1.9.4
zipp==1.0.0
